Refactor for DL-independent channel routing core
