+++ test/test_channels.jl
@@ -465,6 +465,5 @@ end
     end
     # Runs without error
     test_simple()
-    test_fanout()
     test_grid_escape()
 end
